Split test command file-read unselected tests
